Fix ESP-IDF 5.x + Arduino 3.x compatibility
- Change int to int32_t for HA service handler (ESPHome API only has
  template specializations for int32_t, not int)
- Add POSIX compatibility stubs (access, rmdir) for Sqlite3Esp32 library
  which requires these functions that are declared but not implemented
  in ESP-IDF 5.x newlib configuration

+++ components/b48_display_controller/b48_ha_integration.cpp
@@ -62,8 +62,8 @@ void B48HAIntegration::register_services_() {
 
 // --- Service Handlers ---
 
-void B48HAIntegration::handle_delete_message_service_(int message_id) {
-  ESP_LOGD(TAG, "Service b48_delete_message called: message_id=%d", message_id);
+void B48HAIntegration::handle_delete_message_service_(int32_t message_id) {
+  ESP_LOGD(TAG, "Service b48_delete_message called: message_id=%d", (int)message_id);
   if (message_id <= 0) {
     ESP_LOGW(TAG, "Delete message failed: Invalid message_id (%d).", message_id);
     return;
